+++ package.json
@@ -39,50 +39,6 @@
   "workspaces": [
     "packages/*"
   ],
-  "DEPS-NOTE": "Any non-test dependency must be added to the build_package.js script as well",
-  "dependencies": {
-    "@babel/code-frame": "^7.14.5",
-    "@babel/core": "^7.14.8",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-dynamic-import": "^7.14.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.14.5",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
-    "@babel/plugin-proposal-numeric-separator": "^7.14.5",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
-    "@babel/plugin-proposal-private-methods": "^7.14.5",
-    "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
-    "@babel/plugin-syntax-async-generators": "^7.8.4",
-    "@babel/plugin-syntax-json-strings": "^7.8.3",
-    "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
-    "@babel/plugin-syntax-optional-catch-binding": "^7.8.3",
-    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "colors": "^1.4.0",
-    "commander": "^8.2.0",
-    "debug": "^4.1.1",
-    "expect": "^27.2.5",
-    "extract-zip": "^2.0.1",
-    "https-proxy-agent": "^5.0.0",
-    "jest-matcher-utils": "^27.2.5",
-    "jpeg-js": "^0.4.2",
-    "mime": "^2.4.6",
-    "minimatch": "^3.0.3",
-    "ms": "^2.1.2",
-    "open": "^8.2.1",
-    "pirates": "^4.0.1",
-    "pixelmatch": "^5.2.1",
-    "pngjs": "^5.0.0",
-    "progress": "^2.0.3",
-    "proper-lockfile": "^4.1.1",
-    "proxy-from-env": "^1.1.0",
-    "rimraf": "^3.0.2",
-    "source-map-support": "^0.4.18",
-    "stack-utils": "^2.0.3",
-    "ws": "^7.4.6",
-    "yauzl": "^2.10.0",
-    "yazl": "^2.5.1"
-  },
   "devDependencies": {
     "@babel/cli": "^7.14.8",
     "@babel/plugin-transform-typescript": "^7.14.6",
