+++ packages/playwright-test/index.mjs
@@ -13,5 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import playwright from './index.js';
 
-module.exports = require('@playwright/test');
+export const chromium = playwright.chromium;
+export const firefox = playwright.firefox;
+export const webkit = playwright.webkit;
+export const selectors = playwright.selectors;
+export const devices = playwright.devices;
+export const errors = playwright.errors;
+export const _electron = playwright._electron;
+export const _android = playwright._android;
+export const test = playwright.test;
+export const expect = playwright.expect;
+export default playwright.test;
