+++ packages/playwright-test/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@playwright/test",
+  "dependencies": {
+    "@babel/code-frame": "^7.14.5",
+    "@babel/core": "^7.14.8",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-proposal-dynamic-import": "^7.14.5",
+    "@babel/plugin-proposal-export-namespace-from": "^7.14.5",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.5",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
+    "@babel/plugin-proposal-numeric-separator": "^7.14.5",
+    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
+    "@babel/plugin-proposal-private-methods": "^7.14.5",
+    "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
+    "@babel/plugin-syntax-async-generators": "^7.8.4",
+    "@babel/plugin-syntax-json-strings": "^7.8.3",
+    "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
+    "@babel/plugin-syntax-optional-catch-binding": "^7.8.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
+    "@babel/preset-typescript": "^7.14.5",
+    "colors": "^1.4.0",
+    "expect": "^27.2.5",
+    "jest-matcher-utils": "^27.2.5",
+    "minimatch": "^3.0.3",
+    "ms": "^2.1.2",
+    "pirates": "^4.0.1",
+    "pixelmatch": "^5.2.1",
+    "source-map-support": "^0.4.18"
+  },
+  "exports": {
+    ".": {
+      "import": "./index.mjs",
+      "require": "./index.js"
+    },
+    "./": "./"
+  }
+}
