+++ packages/playwright-test/reporter.js
@@ -13,8 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import playwright from './index.js';
 
-export const test = playwright.test;
-export const expect = playwright.expect;
-export default playwright.test;
+// We only export types in reporter.d.ts.
