+++ packages/playwright/package.json
@@ -1,3 +1,5 @@
 {
-  "name": "playwright"
+  "name": "playwright",
+  "dependencies": {
+  }
 }
